[TEST] GetHomeworkContentService 테스트코드 작성
